requirements.txt
Flask>=2.0
requests>=2.28
python-dotenv>=0.21
